refactor: KEEP-356 group Condition node files under lib/workflow/nodes/condition
Pure file relocation. The condition codegen template (lib/codegen-templates/
condition.ts) is intentionally left in place; it is read at runtime by
the workflow download route and will be moved together with all codegen
templates in a later batch.

Moves:
- lib/condition-builder-types.ts -> .../condition/builder-types.ts
- lib/condition-builder-utils.ts -> .../condition/builder-utils.ts
- lib/condition-expression.ts   -> .../condition/expression.ts
- lib/condition-resolver.ts     -> .../condition/resolver.ts
- lib/condition-validator.ts    -> .../condition/validator.ts
- lib/skipped-branch-utils.ts   -> .../condition/skipped-branch.ts
- lib/steps/condition.ts        -> .../condition/step.ts

12 import sites updated. Spec doc reference also updated.
