+++ lib/workflow-codegen.ts
@@ -1,4 +1,4 @@
-import { resolveConditionExpression } from "@/lib/condition-resolver";
+import { resolveConditionExpression } from "@/lib/workflow/nodes/condition/resolver";
 import { buildEdgesBySourceHandle } from "@/lib/edge-handle-utils";
 import { findActionById, flattenConfigFields } from "@/plugins/registry";
 import {
