+++ lib/workflow/nodes/condition/builder-types.ts
