+++ lib/workflow/nodes/condition/validator.ts
